Update cache-handler module and configuration
