+++ frontend/package.json
@@ -63,8 +63,7 @@
   "devDependencies": {
     "@babel/core": "^7.23.7",
     "@faker-js/faker": "^8.3.1",
-    "@neshca/cache-handler": "^0.6.10",
-    "@neshca/json-replacer-reviver": "^1.1.1",
+    "@neshca/cache-handler": "^1.0.4",
     "@storybook/addon-actions": "^7.6.16",
     "@storybook/addon-essentials": "^7.6.16",
     "@storybook/addon-interactions": "^7.6.10",
